Changed all logic to the page
